Create Fitness-Tracking-App
Fitness-Tracking-App is an application allowing these users to be in shape and to manage their diet well for a healthy body. By creating this application, the Student or learner will gain experience and become acquainted with API design and how to link an API to a Mobile application, Or even have knowledge of the Use of Firebase for storing data.