Create Alarm-Clock-Timer
# Alarm-clock-Timer
This is a simple Python program that allows the user to set an alarm for a certain amount of time and choose an alarm sound. The user is prompted to enter the number of minutes and seconds to wait and to choose an alarm sound from a list of four options.

Once the countdown timer reaches zero, the selected alarm sound will play, indicating that the time has elapsed. The user can choose from four different alarm sounds.

This program uses the playsound library to play the selected sound file and the time library to keep track of the elapsed time. The program also utilizes escape sequences to clear the terminal and update the countdown timer display.

# Requirements
This program requires the playsound library to be installed. You can install it using pip: 
pip install playsound

# Usage
To run the program, simply execute the Python script in your preferred environment or terminal:
python alarm_clock.py

Follow the prompts to set the desired time and alarm sound. The program will then start a countdown timer and play the selected sound once the timer reaches zero.